+++ src/app/(authorized)/profile/[id]/page.tsx
@@ -3,6 +3,8 @@ import ProfileServer from '../../../(not_authorized)/(public-info)/public-post-p
 import { redirect } from 'next/navigation';
 import { ROUTES } from '@/appRoutes/routes';
 
+export const dynamic = 'force-dynamic';
+
 const Page = async ({ params }: { params: { id: string } }) => {
   const headersList = headers();
   const idHeaders = headersList.get('id') as string;
