+++ src/app/(authorized)/search/SearchContent/SearchContent.tsx
@@ -6,9 +6,9 @@ import s from './SearchContent.module.scss';
 import { FoundUser } from '../FoundUser';
 import { useDebounce } from '@/utils/useDebaunce';
 import { UserType } from '@/app/(not_authorized)/(public-info)/public-post-page/[id]/types';
-import { getUsers } from '@/app/(authorized)/search/SearchContent/data';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useTranslation } from 'react-i18next';
+import { getUsers } from '@/app/(authorized)/search/SearchContent/actions';
 
 type Props = {
   accessToken: string | null;
@@ -76,7 +76,7 @@ export const SearchContent: React.FC<Props> = ({ accessToken }) => {
     setIsLoad(true);
 
     try {
-      const data = await getUsers(accessToken, {
+      const data = await getUsers({
         search: debouncedSearch,
         pageNumber: pageNumber,
       });
