+++ src/app/(not_authorized)/(public-info)/public-post-page/[id]/PostFix/PostContent/PostContentMobile.tsx
@@ -29,7 +29,7 @@ import { ViewLikesModal } from '@/components/Modals/ViewLikesModal';
 import {
   followToUser,
   unfollowByUser,
-} from '@/app/(authorized)/search/SearchContent/data';
+} from '@/app/(authorized)/search/SearchContent/actions';
 
 type Props = {
   myId?: number;
@@ -50,7 +50,6 @@ export const PostContentMobile = ({
   setEditPost,
   onDeletePost,
   type,
-  token,
 }: Props) => {
   const dispatch = useDispatch();
   const [visiblePopup, setVisiblePopup] = useState(false);
@@ -101,9 +100,7 @@ export const PostContentMobile = ({
   };
 
   const followUnfollow = async (userId: number, isFollowing: boolean) => {
-    isFollowing
-      ? await unfollowByUser(userId, token)
-      : await followToUser(userId, token);
+    isFollowing ? await unfollowByUser(userId) : await followToUser(userId);
   };
 
   const openLikesModal = async () => {
