+++ src/app/(not_authorized)/(public-info)/public-post-page/[id]/ProfileServer/ProfileServer.tsx
@@ -1,3 +1,4 @@
+'use server';
 import { getPosts, getProfile } from '../actions';
 
 import { ProfileInfo } from '../ProfileInfo/ProfileInfo';
