+++ src/components/Modals/SubscribersModal/SubscribersModal.tsx
@@ -12,7 +12,7 @@ import {
   FollowerType,
   GetFollowersDataType,
 } from '@/app/(not_authorized)/(public-info)/public-post-page/[id]/types';
-import { getUserFollowers } from '@/app/(authorized)/search/SearchContent/data';
+import { getUserFollowers } from '@/app/(authorized)/search/SearchContent/actions';
 import { useDebounce } from '@/utils/useDebaunce';
 import { UnsubscribeModal } from '@/components/Modals/UnsubscribeModal';
 
@@ -21,7 +21,6 @@ type Props = {
   userName: string;
   followUnfollow: (userId: number, isFollowing: boolean) => void;
   setShowSubscribersModal: (value: boolean) => void;
-  token: string | null;
 };
 
 export const SubscribersModal = ({
@@ -29,7 +28,6 @@ export const SubscribersModal = ({
   userName,
   followUnfollow,
   setShowSubscribersModal,
-  token,
 }: Props) => {
   const { t } = useTranslation();
   const translate = (key: string): string => t(`MyProfilePage.${key}`);
@@ -49,8 +47,7 @@ export const SubscribersModal = ({
   const fetchUsers = async () => {
     const data: GetFollowersDataType | null = await getUserFollowers(
       userName,
-      debouncedSearch,
-      token
+      debouncedSearch
     );
     if (data) {
       setUsers(data.items);
